fix(sql): widen SUM and INTEGER-FLOAT arithmetic to avoid loss
Review against DuckDB and DataFusion found three numeric defects that the
v0.8.2 tests did not cover.

SUM over INTEGER accumulated in i32, so two rows near i32::MAX overflowed
on ordinary data. PostgreSQL sums int4 into int8 and DuckDB widens to
hugeint for the same reason; SUM(INTEGER) now yields BIGINT.

INTEGER mixed with FLOAT promoted to f32, whose 24-bit mantissa cannot
hold the i32 range, so large integers silently lost magnitude. It now
promotes to DOUBLE. The rule was duplicated between the planner's type
inference and the evaluator, so both were corrected.

Assigning a decimal literal to a FLOAT column was rejected because the
lexer types every decimal literal as DOUBLE and the narrowing was
disallowed, making FLOAT columns impossible to populate. Three separate
compatibility tables encode this rule; all three now permit it.

Refs #61, #67

+++ crates/alopex-sql/src/executor/evaluator/binary_op.rs
@@ -161,8 +161,10 @@ fn numeric_operands(left: &SqlValue, right: &SqlValue) -> Option<NumericOperands
         (Integer(a), BigInt(b)) => Some(NumericOperands::BigInt(i64::from(*a), *b)),
         (BigInt(a), Integer(b)) => Some(NumericOperands::BigInt(*a, i64::from(*b))),
         (BigInt(a), BigInt(b)) => Some(NumericOperands::BigInt(*a, *b)),
-        (Integer(a), Float(b)) => Some(NumericOperands::Float(*a as f32, *b)),
-        (Float(a), Integer(b)) => Some(NumericOperands::Float(*a, *b as f32)),
+        // f32 has 24 bits of mantissa and cannot hold the whole i32 range, so an
+        // INTEGER mixed with FLOAT widens to DOUBLE instead of losing magnitude.
+        (Integer(a), Float(b)) => Some(NumericOperands::Double(f64::from(*a), f64::from(*b))),
+        (Float(a), Integer(b)) => Some(NumericOperands::Double(f64::from(*a), f64::from(*b))),
         (Float(a), Float(b)) => Some(NumericOperands::Float(*a, *b)),
         (BigInt(a), Float(b)) => Some(NumericOperands::Double(*a as f64, f64::from(*b))),
         (Float(a), BigInt(b)) => Some(NumericOperands::Double(f64::from(*a), *b as f64)),
